requirements.txt
streamlit
langchain
langchain-community
PyPDF2
chromadb
sentence-transformers
groq
python-dotenv